configured database connection and selected first query
