+++ app.js
@@ -8,58 +8,13 @@ var cookieParser = require('cookie-parser');
 var bodyParser = require('body-parser');
 
 var routes = require('./routes/index');
-var users = require('./routes/user');
+var groups = require('./routes/groups');
 
 var app = express();
 
 var env = process.env.NODE_ENV || 'development';
 app.locals.ENV = env;
 app.locals.ENV_DEVELOPMENT = env == 'development';
-
-// database connection setup
-var Connection = require('tedious').Connection;
-var config = {
-    userName: 'Nick',
-    password: 'Kinderenvoorkinderen!12',
-    server: 'lvsdb.database.windows.net',
-    // If you are on Microsoft Azure, you need this:
-    options: {encrypt: true, database: 'AdventureWorks'}
-};
-var connection = new Connection(config);
-connection.on('connect', function(err) {
-// If no error, then good to proceed.
-    console.log("Connected");
-    executeStatement();
-});
-
-// execute a query
-var Request = require('tedious').Request;
-var TYPES = require('tedious').TYPES;
-
-function executeStatement() {
-    request = new Request("SELECT * FROM [LVS].[Group]", function(err) {
-        if (err) {
-            console.log(err);}
-    });
-    console.log(request);
-//     var result = "";
-//     // request.on('row', function(columns) {
-//     //     columns.forEach(function(column) {
-//     //       if (column.value === null) {
-//     //         console.log('NULL');
-//     //       } else {
-//     //         result+= column.value + " ";
-//     //       }
-//     //     });
-//     //     console.log(result);
-//     //     result ="";
-//     // });
-
-//     request.on('done', function(rowCount, more) {
-//         console.log(rowCount + ' rows returned');
-//     });
-//     connection.execSql(request);
-}
 
 // view engine setup
 
@@ -76,7 +31,7 @@ app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
 app.use('/', routes);
-app.use('/users', users);
+app.use('/groups', groups);
 
 /// catch 404 and forward to error handler
 app.use(function(req, res, next) {
